refactor: SmallAppInfoColumnView -> AppInfoColumnView로 이름 변경 및 rowSize 추가
